Fixed errors and small css issues
The validation error was fixed by renaming the div elements inside the button to span elements. The small css issue was about the gif that oveflowed on lower window heights. This problem was fixed by not using position absolute and to use flexbox instead to center the gif inside the section element with class ads.